added a unit test to complete a functional test (fixing build)
--HG--
extra : convert_revision : svn%3Aaa2aecf3-ea3e-0410-9d70-716747e7c967/trunk%40427

+++ test/org/mockitousage/verification/DescriptiveMessagesWhenVerificationFailsTest.java
@@ -259,7 +259,7 @@ public class DescriptiveMessagesWhenVerificationFailsTest extends TestBase {
 
     @Test
     public void shouldPrintNullArguments() throws Exception {
-        mock.simpleMethod(null, null);
+        mock.simpleMethod(null, (Integer) null);
         try {
             verify(mock).simpleMethod("test");
             fail();
